Added ability to use button without encoder and allow pin 0 to be used
Added new constructor for button only initialization.  Changes all pin
definitions to int8_t so a negative pin number will disable the button
operation, or also the encoder operation.  There was a previously an
issue with the default button pin being -1, but stored unsigned was pin
255, so the button logic was running unless you explicitly specified pin
0 - this caused pin 0 to be unusable for a button.

Note that if you previously specified pin 0 in your sketch  to disable
the button operation, then this should be changed to -1