Added ability to have multiple buttons on one analog pin/Button race condition fixed
Multiple buttons can be put on one analog pin by defining a range for
the analog input that presents a button press.  Note that this method
only really works for one button pressed at a time.  Multiple buttons
pressed at a time may not register at all, or register as the wrong
button.  A new constructir was added that allowed the analog input range
to be specified.

Also added noInterrupts to getButton to cure race condition