Fixed ambiguous constructor.  Analog buttons are now a seperate object
Removed the ClickEncoder constructor for multiplexed analog buttons.
AnalogButton is now a child class that can be used to defined these
buttons.

Also added a DIgitalButton child class for using a stand alone buttons
(without an encoder).  The ClickEncoder constructor for a stand alone
button still exists, but consider it depricated in favor of
DigitalButton.